Remove dependency on kue.
Kue is causing stability issues with redis and also seems to do too much
for the limited scope of its use in this project. A simple redis based
store in combination with pub sub to communicate with worker should
be sufficient in the short term.

Refactor so that all objects get the entire config object instead of
relavant portions in order not to leak implemntation details all over.